+++ lib/models/api/v1/entieties/tweet_entities.dart
@@ -1,10 +1,10 @@
 import 'dart:convert';
 
-import 'package:tweet_ui/models/api/entieties/hashtag_entity.dart';
-import 'package:tweet_ui/models/api/entieties/media_entity.dart';
-import 'package:tweet_ui/models/api/entieties/mention_entity.dart';
-import 'package:tweet_ui/models/api/entieties/symbol_entity.dart';
-import 'package:tweet_ui/models/api/entieties/url_entity.dart';
+import 'package:tweet_ui/models/api/v1/entieties/hashtag_entity.dart';
+import 'package:tweet_ui/models/api/v1/entieties/media_entity.dart';
+import 'package:tweet_ui/models/api/v1/entieties/mention_entity.dart';
+import 'package:tweet_ui/models/api/v1/entieties/symbol_entity.dart';
+import 'package:tweet_ui/models/api/v1/entieties/url_entity.dart';
 
 /// Provides metadata and additional contextual information about content posted in a tweet.
 class TweetEntities {