+++ lib/models/api/v2/entities/mention_entity_v2.g.dart
@@ -1,0 +1,14 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'mention_entity_v2.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+MentionEntityV2 _$MentionEntityV2FromJson(Map<String, dynamic> json) =>
+    MentionEntityV2(
+      username: json['username'] as String,
+      start: json['start'] as int,
+      end: json['end'] as int,
+    );