+++ lib/src/byline.dart
@@ -24,6 +24,9 @@ class Byline extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final bool createdDateAvailable =
+        tweetVM.getDisplayTweet().createdAt != null;
+
     switch (viewMode) {
       case ViewMode.standard:
         return Column(
@@ -52,20 +55,23 @@ class Byline extends StatelessWidget {
                 ),
               ],
             ),
-            showDate
-                ? Text(
-                    "@" +
-                        tweetVM.getDisplayTweet().userScreenName +
-                        " • " +
-                        tweetVM.getDisplayTweet().createdAt,
-                    textAlign: TextAlign.start,
-                    style: userScreenNameStyle,
-                  )
-                : Text(
-                    "@" + tweetVM.getDisplayTweet().userScreenName,
-                    textAlign: TextAlign.start,
-                    style: userScreenNameStyle,
-                  ),
+            tweetVM.getDisplayTweet().userScreenName.isNotEmpty
+                ? (showDate
+                    ? Text(
+                        "@" +
+                            tweetVM.getDisplayTweet().userScreenName +
+                            (createdDateAvailable
+                                ? " • ${tweetVM.getDisplayTweet().createdAt}"
+                                : ""),
+                        textAlign: TextAlign.start,
+                        style: userScreenNameStyle,
+                      )
+                    : Text(
+                        "@" + tweetVM.getDisplayTweet().userScreenName,
+                        textAlign: TextAlign.start,
+                        style: userScreenNameStyle,
+                      ))
+                : const SizedBox.shrink(),
           ],
         );
       case ViewMode.compact:
@@ -84,36 +90,39 @@ class Byline extends StatelessWidget {
                   mainAxisSize: MainAxisSize.min,
                   mainAxisAlignment: MainAxisAlignment.start,
                   children: <Widget>[
-                    Text(
-                      tweetVM.getDisplayTweet().userName,
-                      style: userNameStyle,
-                      textAlign: TextAlign.start,
-                    ),
+                    if (tweetVM.getDisplayTweet().userName.isNotEmpty)
+                      Text(
+                        tweetVM.getDisplayTweet().userName,
+                        style: userNameStyle,
+                        textAlign: TextAlign.start,
+                      ),
                     Padding(
                       padding: const EdgeInsets.only(left: 2.0),
                       child: VerifiedUsedBadge(
                           tweetVM.getDisplayTweet(), viewMode),
                     ),
-                    Flexible(
-                      child: Padding(
-                        padding: const EdgeInsets.only(left: 4.0),
-                        child: Text(
-                          "@" + tweetVM.getDisplayTweet().userScreenName,
-                          style: userScreenNameStyle,
-                          maxLines: 1,
-                          overflow: TextOverflow.fade,
-                          softWrap: false,
-                          textAlign: TextAlign.start,
+                    if (tweetVM.getDisplayTweet().userScreenName.isNotEmpty)
+                      Flexible(
+                        child: Padding(
+                          padding: const EdgeInsets.only(left: 4.0),
+                          child: Text(
+                            "@" + tweetVM.getDisplayTweet().userScreenName,
+                            style: userScreenNameStyle,
+                            maxLines: 1,
+                            overflow: TextOverflow.fade,
+                            softWrap: false,
+                            textAlign: TextAlign.start,
+                          ),
                         ),
                       ),
-                    ),
                     (showDate == true)
                         ? Flexible(
                             child: Padding(
                               padding:
                                   const EdgeInsets.symmetric(horizontal: 4.0),
                               child: Text(
-                                "• " + tweetVM.getDisplayTweet().createdAt,
+                                "• " +
+                                    (tweetVM.getDisplayTweet().createdAt ?? ""),
                                 style: userScreenNameStyle,
                                 maxLines: 1,
                                 overflow: TextOverflow.fade,